+++ Scripts/GamepadZiplineConnectionController.cs
@@ -135,7 +135,7 @@ namespace ControllerSupport
 
 		public GamepadZiplineConnectionController(InputService inputService, CameraService cameraService,
 			ToolService toolService, PanelTracker panelTracker, ZiplineTowerRegistry ziplineTowerRegistry,
-			KeyBindingRegistry keyBindingRegistry)
+			KeyBindingRegistry keyBindingRegistry, RecentInputDeviceTracker recentInputDeviceTracker)
 		{
 			_inputService = inputService;
 			_cameraService = cameraService;
@@ -144,7 +144,7 @@ namespace ControllerSupport
 			_ziplineTowerRegistry = ziplineTowerRegistry;
 			_keyBindingRegistry = keyBindingRegistry;
 			_confirmGate = new ConfirmReleaseGate(inputService);
-			_handoff = new GamepadMouseHandoff(keyBindingRegistry, inputService);
+			_handoff = new GamepadMouseHandoff(keyBindingRegistry, inputService, recentInputDeviceTracker);
 		}
 
 		public void Load()